keyhive_core: in Document::add_member add delegation to the doc cgka
Problem: agents added to a document with Document::add_member are unable
to decrypt the content of the document because they have not been added
to the Document::cgka. This occurs because documents are groups and so
Document::add_member forwards to Group::add_member_with_manual_content.
add_member_with_manual_content adds the agent to each transitively
included document beneath it, but it doesn't add the agent to the group
itself - because a group is not a document and thus might not have a
cgka.

Solution: Add the agent to the Document::cgka once the
Group::add_member_with_manual_content call completes.

I added a test but I put in a (newly created) integration test in
keyhive_core/tests/encrypt.rs. The keyhive_core/src/keyhive.rs file felt
like it was getting a little unwieldy with all the tests in it. It
should be easier to maintain these tests split across a number of files
in the integration tests.