keyhive_core: update Active::prekey_pairs on prekey add and rotation
Problem: when we rotate or add a prekey, we don't update the
Active::prekey_pairs map, which stores the ShareSecretKey for each
ShareKey. This means that if we receive a CGKA op adding us to a
document using any prekey other than those used at initial creation
time, we don't have the secret key available to do the decryption.

Solution: update the prekey_pairs field on adding or rotating prekeys.
This is achieved by adding a new
`keyhive_core::principal::individual::NewShareKey` type which is
returned by `Individual::expand_prekeys` and
`Individual::rotate_prekey`.