beelay_core: don't generate a contact card on every load
Problem: the code to load Beelay created a keyhive contact card on every
load, incurring a key rotation every time you load the Beelay. This is a
vestige from a previous implementation of the contact card which did not
incur a key rotation but just saved a KeyOp.

Solution: don't create the contact card on load but instead add a
command to beelay to create a new contact card when needed.