beelay-core: Implement session expiry
Problem: So far sessions live forever. Session state is not large, it's
mostly a bunch of pointers to parts of the keyhive, but obviously we
still don't want an unbounded number of them.

Solution: expire sessions after a configurable timeout. This introduces
the question of how to handle requests for sessions which have timed
out. The simplest option is to treat this the same as a request for a
session which does not exist and return an error. This seems like it
would make things harder to debug though, session timeouts will occur
during normal operation and we would probably like clients to retry. On
the other hand, if there is some error on either end and the session ID
just doesn't exist, it would be nice to be able to detect this and bail.
To accomodate this workflow we retain the IDs of expired sessions for 5
minutes and return a SessionExpired error if a client requests an
expired session.