beelay-core: Add a UnixTimestampMillis type (#102)
Problem: We have a UnixTimestamp type which represents time as the
number of seconds since 1970. This type was originally introduced to be
used as the timestamp on messages to avoid replay attacks. In this
context second resolution is great. I have since used this type
elsewhere for other measurements of time where we _do_ need millisecond
precision.

Solution: Add a second UnixTimestampMillis type which represents time as
milliseconds since 1970. The reason to introduce a second type is that
we still want to use a u64 on the wire and second precision is fine for
preventing replay attacks.