keyhive_core: Include agents keys in Keyhive::reachable_prekey_ops_for_agent
Problem: The Keyhive::reachable_prekey_ops_for_agent method is used to
determine what prekey operations should be synchronised with a remote
peer. It currently doesn't return the ops of the requesting agent, which
means that during set reconciliation with a remote peer the remote
always thinks it needs to upload it's own keys.

Solution: Include the keys of the requesting agent in the prekey ops