keyhive_core: notify listeners of received static events
Problem: Keyhive::receive_static_event does not notify the listener of
new events it has received, which means that it's much more complicated
for client code to know whether it needs to persist an event received
over the network.

Solution: notify the listener inside Keyhive::receive_static_event.
Unfortunately this required making several non async methods async. In
particular, it required making `Keyhive::merge` async - i.e.
implementing `MergeAsync` for `Keyhive` instead of `Merge`. It's
actually not possible to implement `MergeAsync` as is for `Keyhive`
because the `MergeAsync::merge_async` method returns a future which is
`Send`, but the `Keyhive` struct is not `Send`, `Keyhive` is also not
`Clone`, which was also required. For now, I have removed the `Send` bound and 
the `Clone` bound, but this obviously makes `MergeAsync` less useful.
However, right now we only use it in places where we don't need `Send`
futures (i.e. we're running on a local runtime), so we can come back to
this later.