beelay_core: Handle saving and loading from multiple archives
Problem: we do not have transactional writes to storage, so we cannot
write to a single archive as we may overwrite concurrent peers writing
to storage.

Solution: write to StorageKey::auth().push("achives").push(<hash of
archive>) so that we don't overwrite existing archives and then load
multiple archives on start. Also implement some logic to remove archives
which we know are now longer needed. Especially on load, if we detect
multiple archives we write a single new archive and remove the original
multiple archives. This means usually we only have to load a single
archive and if there's a crash we clean up on next boot.