keyhive_core: Add an 'ingest_static' feature
Problem: we are currently using the
`Keyhive::ingest_unsorted_static_event` function in `beelay`. To do this
`beelay` depends on the `test_utils` feature of `keyhive_core`, which
also pulls in `proptest`. `proptest` doesn't build for
`wasm32-unknown-unknown`.

Solution: add a new feature called `ingest_static` which is enabled by
default for the `test_utils` feature and only enable `ingest_static` in
`beelay`. The ideal solution would obviously be to implement something
like `ingest_static` properly in `beelay`, but this is a pragmatic
solution for now.