beelay_core: Use custom executor, not futures::executor::LocalExecutor
Problem: the beelay state machine is written using async/await syntax.
To drive the resulting future we use futures::executor::LocalExecutor.
This works but it throws an error if you try and use it inside the
dynamic scope of another executor. For example, if you were to try and
call `beelay_core::Beelay::handle_event` in a task running inside Tokio,
you would get a panic. Additionally, `LocalExecutor` does a whole bunch
of stuff parking and unparking the current thread which we just don't
need to do.

Solution: replace `futures::executor::LocalExecutor` with a custom
executor that just polls a single spawned future to completion. We don't
need to be able to spawn new futures and it is very much not an error to
drive the beelay state machine inside the dynamic scope of another
executor.